requirements.txt
amqp==5.2.0
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
backcall==0.2.0
billiard==4.2.0
blinker==1.7.0
celery==5.3.5
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.3.2
cron-descriptor==1.4.0
decorator==5.1.1
Django==4.2.6
django-celery-beat==2.5.0
django-cors-headers==4.3.0
django-extensions==3.2.3
django-filter==23.3
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
executing==2.0.0
flake8==6.1.0
Flask==3.0.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
kombu==5.3.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.7.0
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pluggy==1.3.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pyflakes==3.1.0
Pygments==2.16.1
PyJWT==2.8.0
pytest==7.4.3
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytils==0.4.1
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
stripe==7.3.0
tomli==2.0.1
traitlets==5.12.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.8
Werkzeug==3.0.1
